+++ SBO-Lib/ChangeLog.md
@@ -12,6 +12,7 @@
   * sboclean: Verify a proper SBO_HOME before trying to clean distfiles.
   * Use color output by default; replace the COLOR setting with NOCOLOR.
   * Util.pm: Account for symlinks when detecting dangerous directory settings.
+  * Solibs.pm, sbocheck: Packages with missing solibs only in /opt are marked; improved disclaimer.
 
 #### 4.2.1 - 2026-07-23
   * sbofind: Show maintainer names unless in CLASSIC mode; search names with --maintainers.